refactor: Update pages fn to folder
Update pages folder:
- export functions
