+++ src/pages/PageHome.jsx
@@ -1,0 +1,24 @@
+import React from 'react';
+/**
+ * #XXX: ESLint
+ * 'React' must be in scope when using JSX
+ */
+
+function PageHome() {
+  // function PageHome() {
+  return (
+    <>
+      <h3>PAGE-Home</h3>
+
+      <hr />
+    </>
+  );
+}
+
+// const PageHome = () => {
+// #NOTE: `Function component is not a function declaration.`
+//   return <h3>Home</h3>;
+// };
+// const PageHome = () => <h3>Home</h3>;
+
+export default PageHome;
