+++ src/pages/PageNotFound.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function PageNotFound() {
+  return (
+    <main>
+      <h3>404</h3>
+      <p>PAGE-NOT-FOUND</p>
+    </main>
+  );
+}
+
+export default PageNotFound;
