+++ src/pages/PageTodos.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function PageTodos() {
+  return (
+    <main>
+      <h3>PAGE-Todos</h3>
+    </main>
+  );
+}
+
+export default PageTodos;
